WORKING: GunStore - Show Compatible optics in weapon
RECOMMENDED UPGRADE
WORKING: Shows compatible optics for the new weapons listed in the gun
store.
The ability to purchase the optic from the ammo list does not work but
the optic does appear. This assists the users in their ability to
identify what optic they can purchase for the weapon.