Add compatible attachment spawning to vehicle loot
Adds new config variable 'A3W_vehicleLootAttachments', when set to a
number greater than zero, it will spawn random but compatible
attachments for each weapon up to the set amount.

The randomAttachments function is also handy in itself, pass a weapon
class and amount, and it will return an array containing up to the
requested amount of random but compatible attachments for the weapon.
Can be used for mission crates, etc.