Working:  Vehicle Lock/Unlock menu actions moved
Community Recommended: Vehicle lock/unlock menu actions now moved to the
bottom of the listing. This is to preserve the original arma menu
actions. This allows players to "get out" or "get in as driver" as top
options.  Players recommended this change because of deaths caused by
the actions being non-standard.